% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sw.R
\name{add_sw}
\alias{add_sw}
\title{Add shortwave radiations to a dataframe}
\usage{
add_sw(x, rad)
}
\arguments{
\item{x}{dataframe}

\item{rad}{dataframe}
}
\value{
x with an additional column sw
}
\description{
Add shortwave radiations to a dataframe
}
\author{
Eva Marques
}
