% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_analysis.R
\name{boxplot_res_lcz}
\alias{boxplot_res_lcz}
\title{Plots boxplot of residuals per Local Climate Zone}
\usage{
boxplot_res_lcz(pro_scores, model = "joint")
}
\description{
Plots boxplot of residuals per Local Climate Zone
}
\author{
Eva Marques
}
