% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_data_bhm.R
\name{data_bhm}
\alias{data_bhm}
\title{Build a data_bhm from a data.frame, data.table, sf or sftime}
\usage{
data_bhm(
  x,
  temp = "temp",
  lat = "lat",
  lon = "lon",
  time = "time",
  build_h,
  build_d,
  dem,
  network
)
}
\arguments{
\item{x}{a data.frame, data.table, sf or sftime}

\item{temp}{the column name for the temperature in degree celcius}

\item{lat}{the column name for the latitude}

\item{lon}{the column name for the longitude}

\item{time}{the column name for the time in UTC}

\item{network}{the name of the network}
}
\value{
a data_bhm object
}
\description{
Build a data_bhm from a data.frame, data.table, sf or sftime
}
\author{
Eva Marques
}
