% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_density_plots.R
\name{density_beta_covar}
\alias{density_beta_covar}
\title{Plot density of beta coefficients of covariates}
\usage{
density_beta_covar(info)
}
\description{
Plot density for beta coefficients of covariates
}
\author{
Eva Marques
}
