% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_density_plots.R
\name{density_beta_obs}
\alias{density_beta_obs}
\title{Plot density for beta per observational source}
\usage{
density_beta_obs(info)
}
\description{
Plot density for beta per observational source
}
\author{
Eva Marques
}
