% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_density_plots.R
\name{density_hyperprec}
\alias{density_hyperprec}
\title{Plot density of hyperprecision}
\usage{
density_hyperprec(mod, info)
}
\description{
Plot density of hyperprecision
}
\author{
Eva Marques
}
