% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_density_plots.R
\name{density_scores}
\alias{density_scores}
\title{Plots density of RMSE and residuals}
\usage{
density_scores(pro_scores, scores)
}
\description{
Plots density of RMSE and residuals
}
\author{
Eva Marques
}
