% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_pred_vs_pro.R
\name{evaluate_pred}
\alias{evaluate_pred}
\title{Evaluate predictions with regards to professional network}
\usage{
evaluate_pred(pred, pro, info, borders)
}
\description{
Evaluate predictions with regards to professional network
}
\author{
Eva Marques
}
