% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_altitude_gradient.R
\name{grad_alt}
\alias{grad_alt}
\title{Correct 2m-temperature with altitude gradient}
\usage{
grad_alt(z, temp)
}
\description{
Correct 2m-temperature with altitude gradient
}
\author{
Eva Marques
}
