% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{project_extent}
\alias{project_extent}
\alias{get_raster_center}
\title{helper functions for UrbanHeat}
\usage{
project_extent(x, from, to)

get_raster_center(filename)
}
\description{
NA
}
\details{
NA
}
\author{
Brandon McNellis
}
