% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_analysis.R
\name{map_median_res}
\alias{map_median_res}
\title{Spatial residual analysis}
\usage{
map_median_res(pro_scores, borders, y_var = "temp_sea", model = "joint")
}
\description{
Map median residuals per professional station
}
\author{
Eva Marques
}
