% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_analysis.R
\name{map_median_res_d_vs_n}
\alias{map_median_res_d_vs_n}
\title{Spatial residual analysis day vs night}
\usage{
map_median_res_d_vs_n(pro_scores, borders)
}
\description{
Map median residuals per professional station
day vs night
}
\author{
Eva Marques
}
