% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_maps.R
\name{map_obs}
\alias{map_obs}
\title{Maps observations from both sources}
\usage{
map_obs(car, cws, ts, borders, y_var = "temp_sea")
}
\description{
Maps observations from both sources
}
\author{
Eva Marques
}
