% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_maps.R
\name{map_pred_sd}
\alias{map_pred_sd}
\title{Map posterior standard deviation}
\usage{
map_pred_sd(pred, borders, model = "joint")
}
\description{
Map posterior standard deviation
}
\author{
Eva Marques
}
