% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_pred_vs_pro.R
\name{overall_scores}
\alias{overall_scores}
\title{Calculate overall scores}
\usage{
overall_scores(pro_scores, y_var = "temp_sea")
}
\description{
Calculate overall scores
}
\author{
Eva Marques
}
