% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_analysis.R
\name{plot_predmean_vs_ref}
\alias{plot_predmean_vs_ref}
\title{Plot posterior mean vs real temperature}
\usage{
plot_predmean_vs_ref(pro_scores, y_var = "temp_sea", model = "joint")
}
\description{
Plot posterior mean vs real temperature
}
\author{
Eva Marques
}
