% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bhm.R
\name{run_bhm}
\alias{run_bhm}
\title{Run Bayesian Hierarchical Model with crowdsourced data}
\usage{
run_bhm(car, cws, pred, ts, sw, temp_reg, borders)
}
\description{
Run Bayesian Hierarchical Model with crowdsourced data
}
\author{
Eva Marques
}
