% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider_charts.R
\name{spider_rmse}
\alias{spider_rmse}
\title{Create spider charts to compare models efficiency within the day}
\usage{
spider_rmse(scores, quantile)
}
\description{
Create spider charts to compare models efficiency within the day
}
\author{
Eva Marques
}
