% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_analysis.R
\name{summarize_pro_scores}
\alias{summarize_pro_scores}
\title{Summarizes scores calculated with professional network}
\usage{
summarize_pro_scores(pro_scores, y_var = "temp_sea")
}
\description{
Summarizes scores calculated with professional network
}
\author{
Eva Marques
}
