% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_plots.R
\name{tiles_fixed_effect_post}
\alias{tiles_fixed_effect_post}
\title{Tileplot for posterior of fixed effects}
\usage{
tiles_fixed_effect_post(scores)
}
\description{
Tileplot for posterior of fixed effects
}
\author{
Eva Marques
}
